ml_service/whisper_service/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
pydub 
 